+++ lectures/06-Harmonic-model/plots-code/spectral-peaks-and-f0.py
@@ -13,7 +13,7 @@ N = 512*2
 M = 511
 t = -60
 w = np.hamming(M)
-start = .8*fs
+start = int(.8*fs)
 hN = N/2
 hM = (M+1)/2
