+++ software/transformations/hpsTransformations.py
@@ -22,13 +22,13 @@ def hpsTimeScale(hfreq, hmag, stocEnv, timeScaling):
 	outFrames = outL*timeScaling[1::2]/maxOutTime          # output time values in frames
 	timeScalingEnv = interp1d(outFrames, inFrames, fill_value=0)    # interpolation function
 	indexes = timeScalingEnv(np.arange(outL))              # generate frame indexes for the output
-	yhfreq = hfreq[round(indexes[0]),:]                    # first output frame
-	yhmag = hmag[round(indexes[0]),:]                      # first output frame
-	ystocEnv = stocEnv[round(indexes[0]),:]                # first output frame
+	yhfreq = hfreq[int(round(indexes[0])),:]                    # first output frame
+	yhmag = hmag[int(round(indexes[0])),:]                      # first output frame
+	ystocEnv = stocEnv[int(round(indexes[0])),:]                # first output frame
 	for l in indexes[1:]:                                  # iterate over all output frame indexes
-		yhfreq = np.vstack((yhfreq, hfreq[round(l),:]))      # get the closest input frame
-		yhmag = np.vstack((yhmag, hmag[round(l),:]))         # get the closest input frame
-		ystocEnv = np.vstack((ystocEnv, stocEnv[round(l),:])) # get the closest input frame
+		yhfreq = np.vstack((yhfreq, hfreq[int(round(l)),:]))      # get the closest input frame
+		yhmag = np.vstack((yhmag, hmag[int(round(l)),:]))         # get the closest input frame
+		ystocEnv = np.vstack((ystocEnv, stocEnv[int(round(l)),:])) # get the closest input frame
 	return yhfreq, yhmag, ystocEnv
 	
 	
@@ -69,13 +69,13 @@ def hpsMorph(hfreq1, hmag1, stocEnv1, hfreq2, hmag2, stocEnv2, hfreqIntp, hmagIn
 	
 	for l in range(L1):                                      # generate morphed frames
 		# identify harmonics that are present in both frames
-		harmonics = np.intersect1d(np.array(np.nonzero(hfreq1[l,:]), dtype=np.int)[0], np.array(np.nonzero(hfreq2[round(L2*l/float(L1)),:]), dtype=np.int)[0])
+		harmonics = np.intersect1d(np.array(np.nonzero(hfreq1[l,:]), dtype=np.int)[0], np.array(np.nonzero(hfreq2[int(round(L2*l/float(L1))),:]), dtype=np.int)[0])
 		# interpolate the frequencies of the existing harmonics
-		yhfreq[l,harmonics] =  (1-hfreqIndexes[l])* hfreq1[l,harmonics] + hfreqIndexes[l] * hfreq2[round(L2*l/float(L1)),harmonics]
+		yhfreq[l,harmonics] =  (1-hfreqIndexes[l])* hfreq1[l,harmonics] + hfreqIndexes[l] * hfreq2[int(round(L2*l/float(L1))),harmonics]
 		# interpolate the magnitudes of the existing harmonics
-		yhmag[l,harmonics] =  (1-hmagIndexes[l])* hmag1[l,harmonics] + hmagIndexes[l] * hmag2[round(L2*l/float(L1)),harmonics]
+		yhmag[l,harmonics] =  (1-hmagIndexes[l])* hmag1[l,harmonics] + hmagIndexes[l] * hmag2[int(round(L2*l/float(L1))),harmonics]
 		# interpolate the stochastic envelopes of both frames
-		ystocEnv[l,:] =  (1-stocIndexes[l])* stocEnv1[l,:] + stocIndexes[l] * stocEnv2[round(L2*l/float(L1)),:]
+		ystocEnv[l,:] =  (1-stocIndexes[l])* stocEnv1[l,:] + stocIndexes[l] * stocEnv2[int(round(L2*l/float(L1))),:]
 	return yhfreq, yhmag, ystocEnv
