+++ software/transformations/stftTransformations.py
@@ -81,7 +81,7 @@ def stftMorph(x1, x2, fs, w1, N1, w2, N2, H1, smoothf, balancef):
 		mX1, pX1 = DFT.dftAnal(x1[pin1-hM1_1:pin1+hM1_2], w1, N1)           # compute dft
 		mX2, pX2 = DFT.dftAnal(x2[pin2-hM2_1:pin2+hM2_2], w2, N2)           # compute dft
 	#-----transformation-----
-		mX2smooth = resample(np.maximum(-200, mX2), mX2.size*smoothf)       # smooth spectrum of second sound
+		mX2smooth = resample(np.maximum(-200, mX2), int(mX2.size*smoothf))       # smooth spectrum of second sound
 		mX2 = resample(mX2smooth, mX1.size)                                 # generate back the same size spectrum
 		mY = balancef * mX2 + (1-balancef) * mX1                            # generate output spectrum
 	#-----synthesis-----
